// Atspm/WebUI/src/features/charts/purduePhaseTermination/types.ts
// #region license
// Copyright 2026 Utah Departement of Transportation
// for WebUI - types.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion
import {
  BaseChartData,
  BaseChartOptions,
  BasePlan,
  ChartType,
} from '@/features/charts/common/types'

export interface PurduePhaseTerminationChartOptions extends BaseChartOptions {
  selectedConsecutiveCount: number
}

export interface PurduePhaseTerminationChartOptionsDefaults {
  selectedConsecutiveCount: { id: number; value: string; option: string }
}

export interface Phase {
  phaseNumber: number
  gapOuts: string[]
  maxOuts: string[]
  forceOffs: string[]
  pedWalkBegins: string[]
  unknownTerminations: string[]
}

export type PurduePhaseTerminationPlan = BasePlan

export interface RawPurduePhaseTerminationData extends BaseChartData {
  consecutiveCount: number
  plans: PurduePhaseTerminationPlan[]
  phases: Phase[]
}

export interface RawPurduePhaseTerminationResponse {
  type: ChartType.PurduePhaseTermination
  data: RawPurduePhaseTerminationData
}